control_plane/requirements.txt
-e .
aws-cdk-core
aws-cdk-aws-iam
aws-cdk-aws-lambda
